change project structure and add environment variable
