+++ pom.xml
@@ -29,6 +29,7 @@
   <properties>
     <java.version>21</java.version>
     <open-api-doc.version>2.0.3</open-api-doc.version>
+    <arrow.version>18.1.0</arrow.version>
     <sonar.organization>antaressimulatorteam</sonar.organization>
     <sonar.projectKey>AntaresSimulatorTeam_antares-datamanager-back</sonar.projectKey>
     <sonar.host.url>https://sonarcloud.io</sonar.host.url>
@@ -162,6 +163,23 @@
       <artifactId>junit-jupiter</artifactId>
       <scope>test</scope>
     </dependency>
+
+    <!--    Timeseries manager -->
+    <dependency>
+      <groupId>org.apache.arrow</groupId>
+      <artifactId>arrow-vector</artifactId>
+      <version>${arrow.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.arrow</groupId>
+      <artifactId>arrow-memory-netty</artifactId>
+      <version>${arrow.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.arrow</groupId>
+      <artifactId>arrow-compression</artifactId>
+      <version>${arrow.version}</version>
+    </dependency>
   </dependencies>
 
   <build>
@@ -246,9 +264,4 @@
       </releases>
     </pluginRepository>
   </pluginRepositories>
-  <modules>
-    <module>timeseries_manager</module>
-  </modules>
-  <packaging>pom</packaging>
-
 </project>
