+++ src/main/java/com/rte_france/antares/datamanager_back/PegaseBackApplication.java
@@ -4,8 +4,8 @@ import com.rte_france.antares.datamanager_back.configuration.CorsConfig;
 import com.rte_france.antares.datamanager_back.configuration.OpenApiConfig;
 import com.rte_france.antares.datamanager_back.configuration.SecurityConfig;
 import com.rte_france.antares.datamanager_back.service.TimeSeriesStorageService;
-import com.rte_france.antares.timeseries_manager.structures.TimeSeriesMatrix;
-import com.rte_france.antares.timeseries_manager.structures.TimeSeriesMatrixColumn;
+import com.rte_france.antares.datamanager_back.util.timeseries_manager.TimeSeriesMatrix;
+import com.rte_france.antares.datamanager_back.util.timeseries_manager.TimeSeriesMatrixColumn;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
