+++ src/main/java/com/rte_france/antares/datamanager_back/configuration/TimeSeriesConfig.java
@@ -1,19 +1,19 @@
 package com.rte_france.antares.datamanager_back.configuration;
 
-import com.rte_france.antares.timeseries_manager.main.ArrowTSReader;
-import com.rte_france.antares.timeseries_manager.main.ArrowTSWriter;
+import com.rte_france.antares.datamanager_back.util.timeseries_manager.TimeSeriesReader;
+import com.rte_france.antares.datamanager_back.util.timeseries_manager.TimeSeriesWriter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class TimeSeriesConfig {
     @Bean
-    public ArrowTSReader arrowTSReader() {
-        return new ArrowTSReader();
+    public TimeSeriesReader arrowTSReader() {
+        return new TimeSeriesReader();
     }
 
     @Bean
-    public ArrowTSWriter arrowTSWriter() {
-        return new ArrowTSWriter();
+    public TimeSeriesWriter arrowTSWriter() {
+        return new TimeSeriesWriter();
     }
 }
