+++ src/main/java/com/rte_france/antares/datamanager_back/util/timeseries_manager/TimeSeriesMatrixColumn.java
@@ -4,13 +4,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.rte_france.antares.timeseries_manager.structures;
+package com.rte_france.antares.datamanager_back.util.timeseries_manager;
 
 import java.util.Arrays;
 import java.util.Objects;
 
 public record TimeSeriesMatrixColumn(String name, double[] values) {
-    public int getSize() {
+    public int size() {
       return values.length;
     }
