+++ src/main/java/com/rte_france/antares/datamanager_back/util/timeseries_manager/TimeSeriesMatrixPrimitive.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.rte_france.antares.timeseries_manager.structures;
+package com.rte_france.antares.datamanager_back.util.timeseries_manager;
 
 /**
  * This class was made for testing purposes with a fully primitive data structure to see if Arrow speeds could be made
